refactor: introduce backend seam in sandbox-core
Split sandbox-core/src into a backend-agnostic core and an isolated Dagger
backend so a second sandbox runtime can be added later. No behavior change.

- backend.rs: SandboxBackend trait + Backend enum (Dagger variant) + the
  RunRequest/ArtifactRequest domain inputs that cross the boundary
- backends/dagger.rs: all dagger_sdk usage — connect/spawn, warm base,
  materialize, execute bodies, SDK error categorization
- session.rs: generic actor (singleton handle, channel, semaphore back-pressure,
  panic recovery) dispatching over Backend; warm-up via prewarm()
- supervisor.rs: in-container supervisor script + supervised_argv + the json
  result contract (parse_supervisor_output)
- validation.rs: path/cwd/pythonpath/snapshot validation + shell_quote
- lib.rs: trimmed to the public API + DTOs + SandboxError; runtime.rs removed

Public NAPI surface (index.d.ts/index.cjs), error codes, DTOs and span names are
unchanged; the only knock-on tweaks are generic-layer error strings that no
longer name Dagger.
