+++ platform/archestra-rs/sandbox-core/src/backends/mod.rs
@@ -1,0 +1,5 @@
+//! concrete sandbox backends. each module owns one runtime's connection,
+//! warm-up, and command-materialisation specifics and implements
+//! `crate::backend::SandboxBackend`.
+
+pub(crate) mod dagger;
